Huge refactoring includes entitys under subdomains also mux adapter.
All entities under subdomains moved to the internal/entity directory.
The reason for this change is preventing redifinition of same entities
(which causes cross domain dependency if a subdomain reaches another
subdomain's entities directly and import cycle sometimes. So I was
gathering the each entity that a subdomain uses under its package which
causes redeclaration of some structures if more than one subdomain
uses same data contained structure.)

The little change on mux router adapter is to prevent route handlers
destroy each other because of the way that mux router wants the route
handler as.