rabbitmq usage methodology and message structs change.
We were using rabbitmq message service for like IncreaseUserPostCount
which is the queue name with one consumer. In a real web application
it is most common and better to use the message queues for event
publishing. So the queue names changed from "increase-user-post-count"
format (which causes a queue that only get consumed for one purpose) to
"post-created" format. So any service need to get notified when a post
is created can consume that queue.